+++ .cspell/custom-dictionary.txt
@@ -143,6 +143,7 @@ GCPVM
 gdamore
 genproto
 gerr
+gochecknoglobals
 gochecknoinit
 gochecknoinits
 goconst
