+++ requirements.txt
@@ -1,11 +1,11 @@
 # This file was autogenerated by uv via the following command:
 #    uv pip compile - -o requirements.txt
-aider-chat==0.63.1
+aider-chat==0.63.2
 aiohappyeyeballs==2.4.3
     # via
     #   aider-chat
     #   aiohttp
-aiohttp==3.10.10
+aiohttp==3.11.2
     # via
     #   aider-chat
     #   huggingface-hub
@@ -161,7 +161,7 @@ jinja2==3.1.4
     #   aider-chat
     #   litellm
     #   torch
-jiter==0.7.0
+jiter==0.7.1
     # via
     #   aider-chat
     #   anthropic
@@ -174,7 +174,7 @@ joblib==1.4.2
     # via
     #   nltk
     #   scikit-learn
-json5==0.9.25
+json5==0.9.28
     # via aider-chat
 jsonschema==4.23.0
     # via
@@ -184,7 +184,7 @@ jsonschema-specifications==2024.10.1
     # via
     #   aider-chat
     #   jsonschema
-litellm==1.51.2
+litellm==1.52.8
     # via aider-chat
 llama-cloud==0.0.6
     # via llama-index-core
@@ -244,12 +244,12 @@ numpy==1.26.4
     #   sentence-transformers
     #   transformers
 ollama==0.3.0
-openai==1.53.0
+openai==1.54.4
     # via
     #   aider-chat
     #   litellm
     #   llama-index-core
-packaging==24.1
+packaging==24.2
     # via
     #   aider-chat
     #   huggingface-hub
@@ -282,6 +282,7 @@ prompt-toolkit==3.0.48
 propcache==0.2.0
     # via
     #   aider-chat
+    #   aiohttp
     #   yarl
 psutil==6.1.0
     # via aider-chat
@@ -350,7 +351,7 @@ referencing==0.35.1
     #   aider-chat
     #   jsonschema
     #   jsonschema-specifications
-regex==2024.9.11
+regex==2024.11.6
     # via
     #   aider-chat
     #   nltk
@@ -368,9 +369,9 @@ requests==2.32.3
     #   tavily-python
     #   tiktoken
     #   transformers
-rich==13.9.3
+rich==13.9.4
     # via aider-chat
-rpds-py==0.20.1
+rpds-py==0.21.0
     # via
     #   aider-chat
     #   jsonschema
@@ -443,7 +444,7 @@ tokenizers==0.19.1
     #   transformers
 torch==2.2.2
     # via sentence-transformers
-tqdm==4.66.6
+tqdm==4.67.0
     # via
     #   aider-chat
     #   huggingface-hub
@@ -503,7 +504,7 @@ yarl==1.17.1
     # via
     #   aider-chat
     #   aiohttp
-zipp==3.20.2
+zipp==3.21.0
     # via
     #   aider-chat
     #   importlib-metadata
